+++ newsletter-bundle/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 4.1.0-RC1 (2015-XX-XX)
 
+ * Require to add a sender address in the channel settings (see contao/core#2896).
  * Throw an exception instead of redirecting to `/contao?act=error` (see contao/core-bundle#395).
  * Adjust the code to be compatible with PHP7 (see contao/core#8018).